image upload in brands

+++ database/migrations/2021_02_10_092321_add_quote_id_to_quotes_item.php
@@ -6,28 +6,15 @@ use Illuminate\Support\Facades\Schema;
 
 class AddQuoteIdToQuotesItem extends Migration
 {
-    /**
-     * Run the migrations.
-     *
-     * @return void
-     */
     public function up()
     {
         Schema::table('quotes_item', function (Blueprint $table) {
-            //
-            $table->float('quotes_id',255);
+            $table->integer('quotes_id');
         });
     }
-
-    /**
-     * Reverse the migrations.
-     *
-     * @return void
-     */
     public function down()
     {
         Schema::table('quotes_item', function (Blueprint $table) {
-            //
         });
     }
 }
